refactor(ravn): address code review feedback on NIU-454 test infrastructure
- Move pytest/pytest-asyncio/pytest-cov/respx from production dependencies
  to [project.optional-dependencies] test extra; update dev extra version pins
- Update make test-ravn to use `uv run --extra test pytest`
- Extract InMemoryChannel, EchoTool, FailingTool to tests/ravn/fixtures/fakes.py
  as the single authoritative definition; both conftest files now import from there
- Import AllowAllPermission/DenyAllPermission from ravn.adapters.permission_adapter
  in both conftest files instead of redefining byte-for-byte duplicates
- Add make_agent() factory to tests/ravn/conftest.py; remove duplicated
  _make_agent helpers from test_agent.py and test_conversation.py
- Remove unused InMemoryChannel.event_types() method
- Fix misleading StrEnum substring assertion in test_agent.py: replace
  `RavnEventType.THOUGHT in ch.events[0].type or any(...)` with
  `any(e.type == RavnEventType.THOUGHT for e in ch.events)`
