+++ Makefile
@@ -105,7 +105,7 @@ test-all: test test-integration test-e2e
 # Ravn-specific tests with coverage
 # --------------------------------------------------------------------------
 test-ravn:
-	uv run pytest tests/ravn/ tests/test_ravn/ -v --tb=short \
+	uv run --extra test pytest tests/ravn/ tests/test_ravn/ -v --tb=short \
 		--cov=src/ravn --cov-report=term-missing --cov-fail-under=85
 
 # --------------------------------------------------------------------------
