+++ pyproject.toml
@@ -19,14 +19,16 @@ dependencies = [
     "websockets>=12.0",
     "python-on-whales>=0.70",
     "gitpython>=3.1",
-    "pytest>=9.0.2",
-    "pytest-asyncio>=1.3.0",
-    "pytest-cov>=7.1.0",
-    "respx>=0.22.0",
     "typer>=0.24.1",
 ]
 
 [project.optional-dependencies]
+test = [
+    "pytest>=9.0.2",
+    "pytest-asyncio>=1.3.0",
+    "pytest-cov>=7.1.0",
+    "respx>=0.22.0",
+]
 rabbitmq = [
     "aio-pika>=9.0",
 ]
@@ -49,14 +51,14 @@ cli = [
     "patchelf>=0.14; sys_platform == 'linux'",
 ]
 dev = [
-    "pytest>=8.0",
-    "pytest-asyncio>=0.23",
-    "pytest-cov>=4.0",
+    "pytest>=9.0.2",
+    "pytest-asyncio>=1.3.0",
+    "pytest-cov>=7.1.0",
     "ruff>=0.8,<0.15",
-    "respx>=0.21",
+    "respx>=0.22.0",
     "cryptography>=42.0",
     "textual>=0.79",
-    "typer>=0.12",
+    "typer>=0.24.1",
 ]
 
 [project.scripts]
