+++ tests/ravn/unit/test_agent.py
@@ -26,32 +26,9 @@ from tests.ravn.conftest import (
     FailingTool,
     InMemoryChannel,
     MockLLM,
+    make_agent,
     make_text_response,
 )
-
-
-def _make_agent(
-    llm: LLMPort,
-    tools=None,
-    *,
-    channel: InMemoryChannel | None = None,
-    permission=None,
-    max_iterations: int = 10,
-) -> tuple[RavnAgent, InMemoryChannel]:
-    ch = channel or InMemoryChannel()
-    perm = permission or AllowAllPermission()
-    agent = RavnAgent(
-        llm=llm,
-        tools=tools or [],
-        channel=ch,
-        permission=perm,
-        system_prompt="You are a test assistant.",
-        model="claude-sonnet-4-6",
-        max_tokens=1024,
-        max_iterations=max_iterations,
-    )
-    return agent, ch
-
 
 # ---------------------------------------------------------------------------
 # Normal (no-tool) flow
@@ -61,7 +38,7 @@ def _make_agent(
 class TestNoToolFlow:
     async def test_single_response(self) -> None:
         llm = MockLLM([make_text_response("Hello!")])
-        agent, _ = _make_agent(llm)
+        agent, _ = make_agent(llm)
 
         result = await agent.run_turn("Hi")
 
@@ -71,7 +48,7 @@ class TestNoToolFlow:
 
     async def test_usage_returned(self) -> None:
         llm = MockLLM([make_text_response("ok", input_tokens=10, output_tokens=5)])
-        agent, _ = _make_agent(llm)
+        agent, _ = make_agent(llm)
 
         result = await agent.run_turn("test")
 
@@ -80,7 +57,7 @@ class TestNoToolFlow:
 
     async def test_session_updated_after_turn(self) -> None:
         llm = MockLLM([make_text_response("response")])
-        agent, _ = _make_agent(llm)
+        agent, _ = make_agent(llm)
 
         await agent.run_turn("Hello")
 
@@ -91,13 +68,11 @@ class TestNoToolFlow:
 
     async def test_channel_receives_thought_and_response_events(self) -> None:
         llm = MockLLM([make_text_response("Hi!")])
-        agent, ch = _make_agent(llm)
+        agent, ch = make_agent(llm)
 
         await agent.run_turn("Hello")
 
-        assert RavnEventType.THOUGHT in ch.events[0].type or any(
-            e.type == RavnEventType.THOUGHT for e in ch.events
-        )
+        assert any(e.type == RavnEventType.THOUGHT for e in ch.events)
         assert any(e.type == RavnEventType.RESPONSE for e in ch.events)
 
     async def test_multi_turn_accumulates_usage(self) -> None:
@@ -107,7 +82,7 @@ class TestNoToolFlow:
                 make_text_response("turn2", input_tokens=15, output_tokens=7),
             ]
         )
-        agent, _ = _make_agent(llm)
+        agent, _ = make_agent(llm)
 
         await agent.run_turn("first")
         await agent.run_turn("second")
@@ -123,7 +98,7 @@ class TestNoToolFlow:
                 make_text_response("answer2"),
             ]
         )
-        agent, _ = _make_agent(llm)
+        agent, _ = make_agent(llm)
 
         await agent.run_turn("q1")
         await agent.run_turn("q2")
@@ -166,7 +141,7 @@ class TestToolCallLoop:
                         usage=TokenUsage(input_tokens=8, output_tokens=3),
                     )
 
-        agent, ch = _make_agent(_CountingLLM(), tools=[tool])
+        agent, ch = make_agent(_CountingLLM(), tools=[tool])
         result = await agent.run_turn("echo ping")
 
         assert result.response == "pong received"
@@ -201,7 +176,7 @@ class TestToolCallLoop:
                         usage=TokenUsage(input_tokens=5, output_tokens=2),
                     )
 
-        agent, ch = _make_agent(_LLM(), tools=[tool])
+        agent, ch = make_agent(_LLM(), tools=[tool])
         await agent.run_turn("go")
 
         assert any(e.type == RavnEventType.TOOL_START for e in ch.events)
@@ -233,7 +208,7 @@ class TestToolCallLoop:
                         usage=TokenUsage(input_tokens=5, output_tokens=2),
                     )
 
-        agent, ch = _make_agent(_LLM(), tools=[])
+        agent, ch = make_agent(_LLM(), tools=[])
         await agent.run_turn("go")
 
         error_results = [e for e in ch.events if e.type == RavnEventType.TOOL_RESULT]
@@ -263,7 +238,7 @@ class TestBudgetExhaustion:
                     usage=TokenUsage(input_tokens=1, output_tokens=1),
                 )
 
-        agent, _ = _make_agent(_InfiniteLLM(), tools=[tool], max_iterations=3)
+        agent, _ = make_agent(_InfiniteLLM(), tools=[tool], max_iterations=3)
 
         with pytest.raises(MaxIterationsError) as exc_info:
             await agent.run_turn("go")
@@ -304,7 +279,7 @@ class TestErrorHandling:
                         usage=TokenUsage(input_tokens=5, output_tokens=2),
                     )
 
-        agent, ch = _make_agent(_LLM(), tools=[tool], permission=DenyAllPermission())
+        agent, ch = make_agent(_LLM(), tools=[tool], permission=DenyAllPermission())
         await agent.run_turn("go")
 
         error_results = [e for e in ch.events if e.type == RavnEventType.TOOL_RESULT]
@@ -337,7 +312,7 @@ class TestErrorHandling:
                         usage=TokenUsage(input_tokens=5, output_tokens=2),
                     )
 
-        agent, _ = _make_agent(_LLM(), tools=[tool])
+        agent, _ = make_agent(_LLM(), tools=[tool])
         result = await agent.run_turn("go")
 
         assert len(result.tool_results) == 1
@@ -372,7 +347,7 @@ class TestErrorHandling:
                         usage=TokenUsage(input_tokens=5, output_tokens=2),
                     )
 
-        agent, _ = _make_agent(_LLM(), tools=[tool])
+        agent, _ = make_agent(_LLM(), tools=[tool])
         result = await agent.run_turn("go")
 
         assert call_count == 2
